Remove KillMode=process
Executing `/var/awslogs/bin/awslogs-agent-launcher.sh` spawns a child process (`/var/awslogs/bin/aws`)

According to the docs, if KillMode is set to `process`, only the main process itself will be killed, its children are kept intact. However, if omitted, the value defaults to `control-group`, which all remaining processes in the control group of this unit will also be killed, including the main process and any child processes.

As we have `logrotate` configured to restart the process on the `postrotate` hook, we must ensure we clean up all child processes that are otherwise left as orphans.